04_StreamingBlueSky/requirements.txt
websockets==12.0
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.1
certifi==2024.2.2
boto3==1.34.69 